+++ internal/helpers/helpers.go
@@ -9,9 +9,9 @@ func GetEnv(key, fallback string) string {
 	return fallback
 }
 
-func Contains(apps []string, app string) bool {
-	for _, a := range apps {
-		if a == app {
+func Contains(slice []string, s string) bool {
+	for _, item := range slice {
+		if item == s {
 			return true
 		}
 	}