+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - e is null issue
+- GetAppList for InMemoryState
 
 ## [0.0.12] - 2022-06-17
 ### Changed